+++ todo-ngx/src/app/app.ts
@@ -1,8 +1,8 @@
-import { Component, inject, Injectable, model, signal } from '@angular/core';
+import { Component, inject, Injectable } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { MatTableModule } from '@angular/material/table';
-import { Category, categorySchema, Todo, todoSchema, TodoState, todoStateSchema } from '../types';
+import { Category, Todo } from '../types';
 import { TodoStatePipe } from './pipes/todo-state-pipe';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -11,20 +11,8 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatSelectModule } from '@angular/material/select';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
-import { MAT_DIALOG_DATA, MatDialog, MatDialogActions, MatDialogClose, MatDialogContent, MatDialogRef, MatDialogTitle } from '@angular/material/dialog';
-import { z } from 'zod';
-
-const todoUpdateFormSchema = z.object({
-  categories: categorySchema.array(),
-  todo: z.object({
-    categoryId: z.uint32().nullable(),
-    title: z.string().max(255),
-    body: z.string(),
-    state: todoStateSchema
-  })
-})
-
-type TodoUpdateForm = z.infer<typeof todoUpdateFormSchema>
+import { MatDialog } from '@angular/material/dialog';
+import { todoUpdateFormSchema, UpdateTodoDialog } from './update-dialog/update-todo-dialog';
 
 @Component({
   selector: 'app-root',
@@ -118,29 +106,5 @@ export class App {
       next: () => this.todos = this.todos.filter((todo) => todo.id !== todoId),
       error: (err) => console.error(err)
     })
-  }
-}
-
-@Component({
-  selector: 'update-todo-dialog',
-  templateUrl: 'update-todo-dialog.html',
-  imports: [
-    MatFormFieldModule,
-    MatInputModule,
-    MatSelectModule,
-    FormsModule,
-    MatDialogTitle,
-    MatDialogContent,
-    MatDialogActions,
-    MatDialogClose,
-  ]
-})
-export class UpdateTodoDialog {
-  readonly dialogRef = inject(MatDialogRef<UpdateTodoDialog>)
-  readonly data = inject<TodoUpdateForm>(MAT_DIALOG_DATA)
-  readonly todo = model(this.data)
-
-  onNoClick(): void {
-    this.dialogRef.close();
   }
 }
