+++ todo-ngx/src/app/update-dialog/update-todo-dialog.html
@@ -27,6 +27,6 @@
   </mat-form-field>
 </mat-dialog-content>
 <mat-dialog-actions>
-  <button matButton (click)="onNoClick()">キャンセル</button>
+  <button matButton (click)="onCancel()">キャンセル</button>
   <button matButton [mat-dialog-close]="todo()" cdkDocusInitial>更新</button>
 </mat-dialog-actions>
